# storage/whatsapp-api-docs/whatsapp-api-swagger-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "WhatsApp API",
        "description": "This is an API documentation for the WhatsApp-like chat application",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/messages": {
            "post": {
                "tags": [
                    "Messages"
                ],
                "summary": "Send a message to a chatroom",
                "operationId": "8078959f5f7a11a644819cf6e68b2edd",
                "parameters": [
                    {
                        "name": "chatroomId",
                        "in": "path",
                        "description": "The ID of the chatroom where the message will be sent",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "message_text",
                                    "chatroom_id"
                                ],
                                "properties": {
                                    "message_text": {
                                        "type": "string",
                                        "example": "Hello, this is a message!"
                                    },
                                    "chatroom_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "attachment": {
                                        "description": "File attachment (optional)",
                                        "type": "string",
                                        "format": "binary",
                                        "example": "path/to/file.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Message sent successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "chatroom_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "sender_user_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "message_text": {
                                            "type": "text",
                                            "example": "Hello, this is a message!"
                                        },
                                        "attachment_path": {
                                            "type": "string",
                                            "example": "pictures/3.jpg"
                                        },
                                        "attachment_type": {
                                            "type": "string",
                                            "example": "image/jpeg"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-11-07T10:00:00Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-11-07T10:00:00Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/messages/{chatroom_id}": {
            "get": {
                "tags": [
                    "Messages"
                ],
                "summary": "List all messages in a chatroom",
                "operationId": "1366a46e1f268860d4a2e49092b05b85",
                "parameters": [
                    {
                        "name": "chatroom_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of messages in a chatroom",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "chatroom_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "sender_user_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "message_text": {
                                                "type": "text",
                                                "example": "Hello, this is a message!"
                                            },
                                            "attachment_path": {
                                                "type": "string",
                                                "example": "pictures/3.jpg"
                                            },
                                            "attachment_type": {
                                                "type": "string",
                                                "example": "image/jpeg"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2024-11-07T10:00:00Z"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2024-11-07T10:00:00Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid chatroom."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
...
}